+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2.x.x (xxxx-xx-xx)
+--
+Bugfixes:
+
+* Search: search page was installed twice.
+
+
 2.6.0 (2011-07-26)
 --
 Bugfixes: